+++ src/main/java/com/phantomwing/eastersdelight/datagen/ModDataMapProvider.java
@@ -20,6 +20,7 @@ public class ModDataMapProvider extends DataMapProvider {
                 // 85% chance
                 .add(ModItems.BOILED_EGG.getId(), new Compostable(0.85f, true), false)
                 .add(ModItems.EGG_SLICE.getId(), new Compostable(0.85f, true), false)
+                .add(ModItems.BUNNY_COOKIE.getId(), new Compostable(0.85f, true), false)
         ;
     }
 }