+++ changelog.md
@@ -1,3 +1,10 @@
+# 0.2.0
+### Additions
+- New food items:
+  - Bunny Cookie
+  - Chocolate Egg
+- Added new configuration options to hide Dyed Egg variations in Creative Mode tab
+
 # 0.1.0
 ### Additions
 - Boiled eggs